Use int for length

+++ Common/Trim.c
@@ -20,7 +20,7 @@ void removeControlChars(unsigned char * const c, size_t * const len) {
 	*len = newLen;
 }
 
-size_t charInvisible(const unsigned char * const c, const size_t len) {
+size_t charInvisible(const unsigned char * const c, const int len) {
 	if (len > 1 && c[0] == 0xCD && c[1] == 0x8F) return 2; // CGJ
 
 	if (len > 2 && (
@@ -32,7 +32,7 @@ size_t charInvisible(const unsigned char * const c, const size_t len) {
 	return 0;
 }
 
-size_t charSpace(const unsigned char * const c, const size_t len) {
+size_t charSpace(const unsigned char * const c, const int len) {
 	if (len > 0 && c[0] == ' ') return 1;
 
 	if (len > 1 && c[0] == 0xC2 && c[1] == 0xA0) return 2; // NBSP
